+++ commons/ava_testsuite/ava_test_suite.go
@@ -12,6 +12,7 @@ func (a AvaTestSuite) GetTests() map[string]testsuite.Test {
 	result["fiveStakingNodeGetValidatorsTest"] = FiveNodeStakingNetworkGetValidatorsTest{}
 	result["fiveStakingNodeFullyConnectedTest"] = FiveNodeStakingNetworkFullyConnectedTest{}
 	result["fiveStakingNodeRpcWorkflowTest"] = FiveNodeStakingNetworkRpcWorkflowTest{}
+	result["fiveStakingNodeDuplicateNodeIdTest"] = FiveNodeStakingNetworkDuplicateIdTest{}
 
 	return result
 }