+++ CHANGELOG.md
@@ -12,12 +12,15 @@
 * Add all five default stakers to staking network bootstrapping
 * Implement test for transferring assets between XChain accounts
 * Implement test for transferring assets from XChain to PChain
+
+### Duplicate Node ID Test
 * Created `GeckoCertProvider` interface that's fed into the `GeckoServiceInitializerCore`, allowing for test writers to customize the certs that the certs get
 * Created two implementations of `GeckoCertProvider`:
     * `StaticGeckoCertProvider`, which provides the exact same predefined cert repeatedly
     * `RandomGeckoCertProvider`, which provides generated certs (optionally, the same random-generated cert each time)
 * Removed `FixedGeckoNetwork` in favor of `TestGeckoNetwork`, which allows for more control over the testnet that gets created
 * Removed the single-node and ten-node Gecko tests; they don't actually test anything useful when compared to the staking network tests
+* Test if the network functions as expected when nodes with duplicate node IDs occur
 
 # 0.2.1
 * Fix tee suppressing exit code of the Docker image